Improve `thrust::cuda_cub::replace` functor handling (#5949)
Reuse functors which exist and proclaim they have copyable arguments when possible
